requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==4.4.1
cffi==1.14.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Django==3.2.6
django-appconf==1.0.4
django-celery==3.3.1
django-celery-email==3.0.0
django-celery-results==2.2.0
djangorestframework==3.12.4
drf-extensions==0.7.1
kombu==4.6.11
librabbitmq==2.0.0
msgpack==1.0.2
pkg_resources==0.0.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
rabbitmq==0.2.0
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
vine==1.3.0
wcwidth==0.2.5

